third_party/blink/web_tests/external/wpt/webvtt/parsing/file-parsing/tests/settings-line-expected.txt

This is a testharness.js-based test.
[FAIL] settings, line
  assert_equals: Failed with cue 31 expected (string) "auto" but got (number) 0.5
Harness: the test ran to completion.
